Refactor transaction validation and script evaluation methods into a new Utils class, update tests to utilize the new structure
