+++ pywaves/txGenerator.py
@@ -378,10 +378,3 @@ class TxGenerator:
 
         return tx
 
-    def validateTx(self, tx):
-        endpoint = f"/debug/validate"
-        return self.pywaves.wrapper(endpoint, tx)
-
-    def evaluateScript(self, address,script):
-        endpoint = f"/utils/script/evaluate/{address}"        
-        return self.pywaves.wrapper(endpoint, script)
