+++ tests/test_Evaluate.py
@@ -3,8 +3,9 @@ from pywaves import address
 import pytest
 import os
 import json
-from pywaves.txGenerator import TxGenerator
+from pywaves.utils import Utils
 from pywaves.txSigner import TxSigner
+from pywaves.txGenerator import TxGenerator
 from tests.helpers import Helpers
 from pywaves import PyWavesException
 
@@ -20,22 +21,24 @@ try:
         assert testwallet is not None        
 
     def test_validateTX_valid():
-        generator = TxGenerator()        
+        txgenerator = TxGenerator()
+        utils = Utils()        
         signer = TxSigner()
-        tx = generator.generateSendWaves(address.Address('3MuqNWyf4RMWz3cqDi4QZRVr9v76LKMjNVZ'), 100000, testwallet.publicKey)    
+        tx = txgenerator.generateSendWaves(address.Address('3MuqNWyf4RMWz3cqDi4QZRVr9v76LKMjNVZ'), 100000, testwallet.publicKey)    
         signer.signTx(tx, testwallet.privateKey)
         print (json.dumps(tx, indent=4))
-        v = generator.validateTx(json.dumps(tx))
+        v = utils.validateTx(json.dumps(tx))
         assert v['valid'] == True
 
     def test_validateTX_invalid():
-        generator = TxGenerator()        
+        utils = Utils()        
+        txgenerator = TxGenerator()
         signer = TxSigner()
-        tx = generator.generateSendWaves(address.Address('3MuqNWyf4RMWz3cqDi4QZRVr9v76LKMjNVZ'), 100000, testwallet.publicKey)    
+        tx = txgenerator.generateSendWaves(address.Address('3MuqNWyf4RMWz3cqDi4QZRVr9v76LKMjNVZ'), 100000, testwallet.publicKey)    
         tx['timestamp'] = tx['timestamp'] - 1000000000000
         signer.signTx(tx, testwallet.privateKey)
         #print (json.dumps(tx, indent=4))
-        v = generator.validateTx(json.dumps(tx))
+        v = utils.validateTx(json.dumps(tx))
         assert v['valid'] == False        
     
     def test_evaluateScript():
@@ -72,11 +75,12 @@ try:
         tx = testwallet.setScript(script, txFee=500000)
         blockchainTx = pw.waitFor(tx['id'])
         
-        generator = TxGenerator()
+        txgenerator = TxGenerator()
+        utils = Utils()
         parameters = [{"type": "string", "value": "test"},
             {"type": "boolean", "value": True}]
 
-        tx = testwallet.txGenerator.generateInvokeScript(
+        tx = txgenerator.generateInvokeScript(
             testwallet.address,
             'storeBooleanValue',
             testwallet.publicKey,
@@ -86,7 +90,7 @@ try:
             txFee=5000000
         )
         print(json.dumps(tx))
-        v = testwallet.txGenerator.evaluateScript(testwallet.address, json.dumps(tx))
+        v = utils.evaluateScript(testwallet.address, json.dumps(tx))
         print(v)
         assert v.get('complexity', None) == 2
